+++ bluebottle/events/effects.py
@@ -1,95 +1,32 @@
-from datetime import timedelta
-from django.utils import timezone
+from django.utils.translation import ugettext_lazy as _
 
-from bluebottle.activities.triggers import Complete
-from bluebottle.fsm.effects import TransitionEffect, RelatedTransitionEffect
-from bluebottle.notifications.effects import NotificationEffect
-from bluebottle.fsm.triggers import ModelChangedTrigger, ModelDeletedTrigger
-
-from bluebottle.events.models import Event, Participant
-from bluebottle.events.messages import EventDateChanged
-from bluebottle.events.states import EventStateMachine, ParticipantStateMachine
+from bluebottle.fsm.effects import Effect
 
 
-class CapacityChanged(ModelChangedTrigger):
-    field = 'capacity'
+class SetTimeSpent(Effect):
+    "Set time spent on participants"
+    post_save = False
 
-    effects = [
-        TransitionEffect('unfill', conditions=[EventStateMachine.is_not_full]),
-        TransitionEffect('fill', conditions=[EventStateMachine.is_full]),
-    ]
+    def execute(self, **kwargs):
+        if not self.instance.time_spent:
+            self.instance.time_spent = self.instance.activity.duration
 
-
-class DateChanged(ModelChangedTrigger):
-    field = 'start'
-
-    effects = [
-        NotificationEffect(EventDateChanged),
-        TransitionEffect('succeed', conditions=[EventStateMachine.should_finish, EventStateMachine.has_participants]),
-        TransitionEffect('close', conditions=[EventStateMachine.should_finish, EventStateMachine.has_no_participants]),
-        TransitionEffect('reopen', conditions=[EventStateMachine.should_open]),
-        TransitionEffect('fill', conditions=[EventStateMachine.is_full]),
-    ]
-
-
-class Started(ModelChangedTrigger):
-    @property
-    def is_valid(self):
-        "The event has started"
-        return (
-            self.instance.duration and
-            (self.instance.start and self.instance.start < timezone.now()) and
-            self.instance.status not in ('succeeded', 'closed', )
+    def __unicode__(self):
+        return _('Set time spent to {} on {}').format(
+            self.instance.activity.duration,
+            self.instance
         )
 
-    effects = [
-        TransitionEffect('start', conditions=[EventStateMachine.should_start, EventStateMachine.has_participants]),
-    ]
 
+class ResetTimeSpent(Effect):
+    "Set time spent to 0 if it was not overriden"
+    post_save = False
 
-class Finished(ModelChangedTrigger):
-    @property
-    def is_valid(self):
-        "The event has ended"
-        return (
-            self.instance.duration and
-            (
-                self.instance.start and
-                self.instance.start + timedelta(hours=self.instance.duration) < timezone.now()
-            ) and
-            self.instance.status not in ('succeeded', 'closed', )
+    def execute(self, **kwargs):
+        if self.instance.time_spent == self.instance.activity.duration:
+            self.instance.time_spent = 0
+
+    def __unicode__(self):
+        return _('Reset time spent to 0 on {}').format(
+            self.instance
         )
-
-    effects = [
-        TransitionEffect('succeed', conditions=[EventStateMachine.should_finish, EventStateMachine.has_participants]),
-        TransitionEffect('close', conditions=[EventStateMachine.should_finish, EventStateMachine.has_no_participants]),
-    ]
-
-
-Event.triggers = [Complete, CapacityChanged, DateChanged, Started, Finished]
-
-
-class ParticipantDeleted(ModelDeletedTrigger):
-    field = 'start'
-
-    effects = [
-        RelatedTransitionEffect(
-            'activity',
-            'close',
-            conditions=[
-                ParticipantStateMachine.event_is_finished,
-                ParticipantStateMachine.event_will_be_empty
-            ]
-        ),
-        RelatedTransitionEffect(
-            'activity',
-            'unfill',
-            conditions=[
-                ParticipantStateMachine.event_will_become_open,
-                ParticipantStateMachine.event_is_not_finished
-            ],
-        ),
-    ]
-
-
-Participant.triggers = [ParticipantDeleted]
