+++ bluebottle/events/states.py
@@ -1,18 +1,8 @@
-from django.utils.translation import ugettext_lazy as _
 from django.utils import timezone
+from django.utils.translation import ugettext_lazy as _
 
-from bluebottle.fsm.state import State, EmptyState, Transition
-from bluebottle.fsm.effects import (
-    TransitionEffect,
-    RelatedTransitionEffect,
-    Effect
-)
-from bluebottle.follow.effects import (
-    FollowActivityEffect, UnFollowActivityEffect
-)
 from bluebottle.activities.states import ActivityStateMachine, ContributionStateMachine
-
-from bluebottle.events.models import Event, Participant
+from bluebottle.events.effects import SetTimeSpent, ResetTimeSpent
 from bluebottle.events.messages import (
     EventSucceededOwnerMessage,
     EventClosedOwnerMessage,
@@ -20,37 +10,16 @@ from bluebottle.events.messages import (
     ParticipantApplicationMessage,
     ParticipantApplicationManagerMessage,
 )
-
+from bluebottle.events.models import Event, Participant
+from bluebottle.follow.effects import (
+    FollowActivityEffect, UnFollowActivityEffect
+)
+from bluebottle.fsm.effects import (
+    TransitionEffect,
+    RelatedTransitionEffect
+)
+from bluebottle.fsm.state import State, EmptyState, Transition
 from bluebottle.notifications.effects import NotificationEffect
-
-
-class SetTimeSpent(Effect):
-    "Set time spent on participants"
-    post_save = False
-
-    def execute(self, **kwargs):
-        if not self.instance.time_spent:
-            self.instance.time_spent = self.instance.activity.duration
-
-    def __unicode__(self):
-        return _('Set time spent to {} on {}').format(
-            self.instance.activity.duration,
-            self.instance
-        )
-
-
-class ResetTimeSpent(Effect):
-    "Set time spent to 0 if it was not overriden"
-    post_save = False
-
-    def execute(self, **kwargs):
-        if self.instance.time_spent == self.instance.activity.duration:
-            self.instance.time_spent = 0
-
-    def __unicode__(self):
-        return _('Reset time spent to 0 on {}').format(
-            self.instance
-        )
 
 
 class EventStateMachine(ActivityStateMachine):
